// src/server/shared/lib/pncp/models/IncluirPlanoContratacaoItemDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type IncluirPlanoContratacaoItemDTO = {
  numeroItem: number;
  categoriaItemPca: number;
  descricao?: string;
  unidadeFornecimento: string;
  quantidade: number;
  valorUnitario: number;
  valorTotal: number;
  valorOrcamentoExercicio: number;
  unidadeRequisitante?: string;
  dataDesejada: string;
  grupoContratacaoCodigo?: string;
  grupoContratacaoNome?: string;
  catalogo: number;
  classificacaoCatalogo: '1' | '2';
  codigoItem?: string;
  classificacaoSuperiorCodigo: string;
  classificacaoSuperiorNome: string;
  pdmCodigo?: string;
  pdmDescricao?: string;
};