// src/server/shared/lib/pncp/models/IRecuperarPlanoItemDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type IRecuperarPlanoItemDTO = {
  pdmCodigo?: string;
  dataAtualizacao?: string;
  cnpj?: string;
  dataInclusao?: string;
  numeroItem?: number;
  dataPublicacaoPncp?: string;
  codigoUnidade?: string;
  nomeUnidade?: string;
  descricao?: string;
  anoPca?: number;
  valorTotal?: number;
  sequencialPca?: number;
  pdmDescricao?: string;
  codigoItem?: string;
  unidadeRequisitante?: string;
  grupoContratacaoCodigo?: string;
  grupoContratacaoNome?: string;
  classificacaoSuperiorCodigo?: string;
  classificacaoSuperiorNome?: string;
  unidadeFornecimento?: string;
  quantidade?: number;
  valorUnitario?: number;
  valorOrcamentoExercicio?: number;
  dataDesejada?: string;
  classificacaoCatalogoId?: number;
  nomeCatalogo?: string;
  categoriaItemPcaid?: number;
  nomeClassificacao?: string;
  catalogoId?: number;
  categoriaItemPcaNome?: string;
};