// src/server/shared/lib/pncp/models/PlanoContratacaoComItensDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { IRecuperarPlanoItemDTO } from './IRecuperarPlanoItemDTO';
export type PlanoContratacaoComItensDTO = {
  itens?: Array<IRecuperarPlanoItemDTO>;
  unidadeResponsavel?: string;
  uasg?: string;
};