// src/server/shared/lib/pncp/models/PlanoContratacaoItemOrgaoToDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PlanoContratacaoItemOrgaoToDTO = {
  dataAtualizacao?: string;
  cnpj?: string;
  dataPublicacaoPncp?: string;
  anoPca?: number;
  valorTotal?: number;
  quantidade?: number;
  poder?: string;
  esfera?: string;
  razaoSocial?: string;
};