// src/server/shared/lib/pncp/models/PlanoContratacaoOrgaoDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PlanoContratacaoOrgaoDTO = {
  dataPublicacaoPncp?: string;
  codigoUnidade?: string;
  nomeUnidade?: string;
  valorTotal?: number;
  cnpj?: string;
  anoPca?: number;
  razaoSocial?: string;
  dataAtualizacao?: string;
  dataAtualizacaoGlobalPCA?: string;
  esfera?: string;
  poder?: string;
  sequencialPca?: number;
  quantidade?: number;
  numeroControlePNCP?: string;
};