// src/server/shared/lib/pncp/models/PlanoSequencialConsolidadoDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type PlanoSequencialConsolidadoDTO = {
  valorTotal?: number;
  quantidade?: number;
  dataAtualizacao?: string;
  dataAtualizacaoGlobalPCA?: string;
  dataPublicacaoPncp?: string;
  codigoUnidade?: string;
  nomeUnidade?: string;
  anoPca?: number;
  cnpj?: string;
  razaoSocial?: string;
  numeroControlePNCP?: string;
  usuario?: string;
  sequencialPca?: number;
  municipio?: string;
  uf?: string;
};