// src/server/shared/lib/pncp/models/RetificarParcialPlanoContratacaoItemDTO.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RetificarParcialPlanoContratacaoItemDTO = {
  categoriaItemPca?: number;
  catalogo?: number;
  classificacaoSuperiorCodigo?: string;
  classificacaoSuperiorNome?: string;
  pdmCodigo?: string;
  pdmDescricao?: string;
  classificacaoCatalogo?: '1' | '2';
  codigoItem?: string;
  descricao?: string;
  unidadeFornecimento?: string;
  quantidade?: number;
  valorUnitario?: number;
  valorTotal?: number;
  valorOrcamentoExercicio?: number;
  unidadeRequisitante?: string;
  dataDesejada?: string;
  grupoContratacaoCodigo?: string;
  grupoContratacaoNome?: string;
  justificativa?: string;
  numeroItem?: number;
};